requirements.txt
bittensor==7.3.1
python-telegram-bot==13.13
python-dotenv
python-dateutil
torch
requests
pandas
